feat: Create application on apps deploy
