+++ cmd/meroxa/root/apps/apps.go
@@ -64,7 +64,6 @@ func (*Apps) FeatureFlag() (string, error) {
 
 func (*Apps) SubCommands() []*cobra.Command {
 	return []*cobra.Command{
-		builder.BuildCobraCommand(&Create{}),
 		builder.BuildCobraCommand(&Describe{}),
 		builder.BuildCobraCommand(&Deploy{}),
 		builder.BuildCobraCommand(&Init{}),
