+++ cmd/meroxa/root/apps/deploy_test.go
@@ -1,7 +1,19 @@
 package apps
 
 import (
+	"context"
+	"encoding/json"
+	"fmt"
+	"os"
+	"reflect"
 	"testing"
+
+	"github.com/meroxa/meroxa-go/pkg/meroxa"
+	turbine "github.com/meroxa/turbine/init"
+
+	"github.com/golang/mock/gomock"
+	"github.com/meroxa/cli/log"
+	"github.com/meroxa/meroxa-go/pkg/mock"
 
 	"github.com/meroxa/cli/cmd/meroxa/builder"
 	"github.com/meroxa/cli/utils"
@@ -41,4 +53,92 @@ func TestDeployAppFlags(t *testing.T) {
 			}
 		}
 	}
+}
+
+const tempAppDir = "test-app"
+
+func TestCreateApplication(t *testing.T) {
+	ctx := context.Background()
+	ctrl := gomock.NewController(t)
+	client := mock.NewMockClient(ctrl)
+	logger := log.NewTestLogger()
+	name := "my-application"
+	lang := GoLang
+
+	// Create application locally
+	path, err := initLocalApp(name)
+	if err != nil {
+		t.Fatalf("not expected error, got \"%s\"", err.Error())
+	}
+	defer func() {
+		_ = os.RemoveAll(tempAppDir)
+	}()
+
+	ai := &meroxa.CreateApplicationInput{
+		Name:     name,
+		Language: lang,
+	}
+
+	a := &meroxa.Application{
+		Name:     name,
+		Language: lang,
+	}
+
+	client.
+		EXPECT().
+		CreateApplication(
+			ctx,
+			ai,
+		).
+		Return(a, nil)
+
+	d := &Deploy{
+		client: client,
+		logger: logger,
+		path:   path,
+		lang:   lang,
+	}
+
+	err = d.createApplication(ctx)
+
+	if err != nil {
+		t.Fatalf("not expected error, got \"%s\"", err.Error())
+	}
+
+	gotLeveledOutput := logger.LeveledOutput()
+	wantLeveledOutput := fmt.Sprintf(`Creating application %q with language %q...
+Application %q successfully created!
+`, name, lang, name)
+
+	if gotLeveledOutput != wantLeveledOutput {
+		t.Fatalf("expected output:\n%s\ngot:\n%s", wantLeveledOutput, gotLeveledOutput)
+	}
+
+	gotJSONOutput := logger.JSONOutput()
+	var gotApplication meroxa.Application
+	err = json.Unmarshal([]byte(gotJSONOutput), &gotApplication)
+	if err != nil {
+		t.Fatalf("not expected error, got %q", err.Error())
+	}
+
+	if !reflect.DeepEqual(gotApplication, *a) {
+		t.Fatalf("expected \"%v\", got \"%v\"", *a, gotApplication)
+	}
+}
+
+func initLocalApp(name string) (string, error) {
+	if err := os.Mkdir(tempAppDir, 0700); err != nil {
+		return "", err
+	}
+
+	if err := turbine.Init(name, tempAppDir); err != nil {
+		return "", err
+	}
+
+	cwd, err := os.Getwd()
+	if err != nil {
+		return "", err
+	}
+
+	return fmt.Sprintf("%s/%s/%s", cwd, tempAppDir, name), nil
 }
